modules/socket: Added support for IPv6
This changelist adds support for IPv6 which is limited to what
is supported by lwip.

The main lwip limitations are:
- DHCPv6 only supports stateless mode
  - Trying to enable DHCPv6 stateless mode will return an error
- Static IPv6 configuration does not support gateway address or prefix length
  - If these options are configured they will not be used
  - lwip currently assumes all prefix lengths are /64
  - lwip currently only supports routes learned by NDP

Currently IPv6 Duplicate Address Detection (DAD) is disabled
because it causes issues with AAT because addresses come
up in tentative state and are not immediately usable.
In order to enable DAD there needs to be an API so AAT tests
can wait for DAD to complete where necessary.