modules/packetio: Added prefetch and refactor interface dispatch code
Prefetch was added when accessing the packet data to
lookup the interface.

Access to the packet data in the interface dispatch was
minimized by reorganizing code to use non-null interface
to indicate unicast and nullptr to indicate multicast.

In order to support this, the interface packet processing
was also reworked so that unicast packets which do not match an
interface are filtered out and packets are filtered/processed
in a way to reduce the number of interface lookups required.