lib/packet: add default values to header fields
Add the ability to set default values in header definitions. When
present, the code generators will add an additional setter function to
set all of the default fields. Additionally, the transmogrify functions
will call the default setter when instantiating protocol headers.
Finally, propagate the default values to the swagger specification so
they are visible to clients.

Also, fix a self correcting error in the make_protocol_header scripts
while here.