workers/packetio: box transmit table entries
The immer containers provide a `box` entry to allow storing pointers
instead of values in the immer tables. This can allow higher object
density in the immer containers if the stored objects are larger than a
pointer.

Update the transmit table to use boxes, as this reduces the number of
stalled cycles when benchmarking.