packet/generator: add stream id to tx flows
The Spirent signature stream id property was not set on tx flow results
for Spirent signature flows, so set it.

Update the sequence code to set the stream id in the signature
configuraton passed to the source object. This prevents the source from
having to calculate the value, which was incorrect when multiple
definitions were configured.

Note that Spirent stream id's consist of a 32 bit value with a 16 bit
stream id in the upper bits and a 16 bit flow id in the lower bits.
Additionally, flow id 0 is reserved so offset the flow index by 1.