circleci: put hard limits on parallel compilation
The kernel Out Of Memory (OOM) killer is occasionally killing build
commands when we execute in some containers.  Apparently, the amound of
memory available is a hard limit, but the number of cores available is
not.  We hit the memory limit when we build in a 36 core container with
only 4GB of usable memory.